pyproject.toml
[project]
name = "agentic-rag-backend"
version = "0.1.0"
description = "Agentic RAG backend with LangGraph, FastAPI, and Langfuse observability"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "langchain>=0.3",
    "langchain-openai>=0.3",
    "langgraph>=0.3",
    "langfuse>=3",
    "pydantic>=2",
    "pydantic-settings>=2",
    "python-dotenv>=1",
    "sse-starlette>=2",
    "langgraph-checkpoint-postgres>=3",
    "psycopg[binary]>=3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.24",
    "httpx>=0.27",
    "ruff>=0.8",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
